:sparkles: :gear: Renommage de treat en parse; reorg
Ce commit n'est pas complet, mais l'idée est lisible
